fastrlap-mujoco-env/requirements.txt
appdirs==1.4.4
attrs==22.2.0
cached-property==1.5.2
cffi==1.15.1
charset-normalizer==3.1.0
chex==0.1.7
click==8.1.3
cloudpickle==2.2.1
contextlib2==21.6.0
Cython==0.29.34
decorator==4.4.2
dm-control==1.0.9
dm-env==1.6
dm-robotics-agentflow==0.5.0
dm-robotics-controllers==0.5.0
dm-robotics-geometry==0.5.0
dm-robotics-manipulation==0.5.0
dm-robotics-moma==0.5.0
dm-robotics-transformations==0.5.0
dm-tree==0.1.8
docker-pycreds==0.4.0
etils==1.2.0
fasteners==0.18
flax==0.6.8
gast==0.5.3
gitdb==4.0.10
GitPython==3.1.31
glfw==2.5.9
idna==3.4
imageio==2.27.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.3.0
importlib-resources==5.12.0
labmaze==1.0.6
lockfile==0.12.2
lxml==4.9.2
markdown-it-py==2.2.0
mdurl==0.1.2
ml-collections==0.1.1
ml-dtypes==0.1.0
moviepy==1.0.3
msgpack==1.0.5
mujoco==2.3.1.post1
mujoco-py==2.1.2.14
nest-asyncio==1.5.6
numpy==1.24.2
numpy-quaternion==2022.4.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-nvcc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.9.0.131
nvidia-cufft-cu11==10.9.0.58
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
opencv-python==4.6.0.66
opt-einsum==3.3.0
optax==0.1.4
orbax==0.1.7
pathtools==0.1.2
Pillow==9.5.0
proglog==0.1.10
protobuf==4.22.3
psutil==5.9.4
pycparser==2.21
pydot==1.4.2
Pygments==2.15.0
PyOpenGL==3.1.6
pyparsing==3.0.0
PyYAML==6.0
requests==2.28.2
rich==13.3.4
scipy==1.10.1
sentry-sdk==1.19.1
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
tensorflow-probability==0.19.0
tensorstore==0.1.35
toolz==0.12.0
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
wandb==0.14.2
zipp==3.15.0